feat: instrument the martech stack
